Split up app into routers and test with Cypress
