+++ package.json
@@ -4,14 +4,17 @@
     "axios": "^1.1.0",
     "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
+    "cypress": "^11.2.0",
     "docker": "^1.0.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "jest": "^29.3.1",
     "jsdom": "^20.0.1",
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^3.1.2",
     "nodemon": "^2.0.20",
     "prisma": "^4.5.0",
+    "supertest": "^6.3.2",
     "unique-random-array": "^3.0.0",
     "zod": "^3.19.1"
   },
@@ -20,13 +23,15 @@
   "description": "scraping",
   "main": "programmers.js",
   "scripts": {
-    "test": "start",
-    "dev": "npx nodemon programmers.js",
-    "start": "node programmers.js",
+    "test": "npx jest",
+    "test:watch": "npx jest --watch",
+    "dev": "npx nodemon app.js",
+    "start": "node app.js",
     "scrape": "node scraper.js",
     "prisma:seed": "node ./prisma/seeders/programmers.js",
     "prisma:migrate:deploy": "npx prisma migrate deploy",
-    "build": "npm run prisma:migrate:deploy"
+    "build": "npm run prisma:migrate:deploy",
+    "cypress:open": "cypress open"
   },
   "author": "karin hogenbirk",
   "license": "ISC"
