+++ public/login.html
@@ -68,7 +68,7 @@ form.addEventListener('submit', function(event) {
 
 //send data to post
     async function submitForm(email, password) {       
-      const response = await fetch("./auth/login", {
+      const response = await fetch("./programmers/login", {
         method: "POST",
         body: JSON.stringify({ email: email, password: password }),
         headers: {
